+++ frontend/muvie/src/atoms/SearchList.vue
@@ -59,15 +59,14 @@ export default {
   data() {
     return {
       showButtons: false,
-      // videoId: null,
     };
   },
   props: {
     item: Object,
   },
   methods: {
-    addPlaylist(fwg) {
-      this.$store.dispatch("addPlaylist", fwg.id);
+    addPlaylist(item) {
+      this.$store.dispatch("addPlaylist", item.id);
     },
     truncateTitle(title) {
       const maxLength = 42;
@@ -81,10 +80,7 @@ export default {
         title: item.title,
         artist: item.artist,
       });
-      // this.$store.dispatch("getVideoDuration", this.videoId);
-      // this.$store.dispatch("addPlaylist", item.id);
-
-      // console.log("재생은 아직");
+      this.$store.dispatch("addPlaylist", item.id);
     },
     like(item) {
       this.$store.dispatch("like", item.id);
@@ -93,12 +89,6 @@ export default {
       this.$store.dispatch("unlike", item.id);
     },
   },
-  // computed: {
-  //   ...mapState({
-  //     // videoId를 getters를 통해 가져옴
-  //     videoId: (state) => state.mypageStore.getters.getVideoId,
-  //   }),
-  // },
 };
 </script>
