+++ frontend/muvie/src/store/modules/mypageStore.js
@@ -81,38 +81,38 @@ const mypageStore = {
         commit("updateLikeState", res.like_list);
       });
     },
-    playMusic({ commit }, { title, artist }) {
-      console.log(title, artist);
-      const searchRequest = axios.get(
-        `https://content-youtube.googleapis.com/youtube/v3/search?q=${title}${artist}&part=snippet&maxResults=1&type=video&key=${process.env.VUE_APP_YOUTUBE_APIKEY}`
-      );
+    // playMusic({ commit }, { title, artist }) {
+    //   console.log(title, artist);
+    //   const searchRequest = axios.get(
+    //     `https://content-youtube.googleapis.com/youtube/v3/search?q=${title}${artist}&part=snippet&maxResults=1&type=video&key=${process.env.VUE_APP_YOUTUBE_APIKEY}`
+    //   );
 
-      searchRequest
-        .then((res) => {
-          console.log("비디오 아이디");
-          const { videoId } = res.data.items[0].id;
-          console.log(videoId);
-          commit("setVideoId", videoId);
+    //   searchRequest
+    //     .then((res) => {
+    //       console.log("비디오 아이디");
+    //       const { videoId } = res.data.items[0].id;
+    //       console.log(videoId);
+    //       commit("setVideoId", videoId);
 
-          const durationRequest = axios.get(
-            `https://www.googleapis.com/youtube/v3/videos?part=contentDetails&maxResults=1&id=${videoId}&key=${process.env.VUE_APP_YOUTUBE_APIKEY}`
-          );
+    //       const durationRequest = axios.get(
+    //         `https://www.googleapis.com/youtube/v3/videos?part=contentDetails&maxResults=1&id=${videoId}&key=${process.env.VUE_APP_YOUTUBE_APIKEY}`
+    //       );
 
-          durationRequest
-            // eslint-disable-next-line no-shadow
-            .then((res) => {
-              const { duration } = res.data.items[0].contentDetails;
-              console.log(duration);
-              commit("setDuration", duration);
-            })
-            .catch((error) => {
-              console.error("API 요청 에러:", error);
-            });
-        })
-        .catch((error) => {
-          console.error("API 요청 에러:", error);
-        });
-    },
+    //       durationRequest
+    //         // eslint-disable-next-line no-shadow
+    //         .then((res) => {
+    //           const { duration } = res.data.items[0].contentDetails;
+    //           console.log(duration);
+    //           commit("setDuration", duration);
+    //         })
+    //         .catch((error) => {
+    //           console.error("API 요청 에러:", error);
+    //         });
+    //     })
+    //     .catch((error) => {
+    //       console.error("API 요청 에러:", error);
+    //     });
+    // },
     // playMusic({ commit }, { title, artist }) {
     //   console.log(title, artist);
     //   axios
@@ -144,6 +144,7 @@ const mypageStore = {
     // },
     addPlaylist({ commit }, id) {
       axios.post(`${BASE_URL}/music/playlist/${id}`).then((res) => {
+        console.log("플레이 리스트 추가");
         console.log(res);
         commit("updatePlayListState");
       });
@@ -217,11 +218,11 @@ const mypageStore = {
 
       console.log("마이페이지 성공");
     },
-    getUserPlayList(state, payload) {
-      state.play_list = payload.play_list;
-      console.log(state.play_list);
-      console.log("재생목록 요청 성공 뿌뿌");
-    },
+    // getUserPlayList(state, payload) {
+    //   state.play_list = payload.play_list;
+    //   console.log(state.play_list);
+    //   console.log("재생목록 요청 성공 뿌뿌");
+    // },
     getPhoto(state, payload) {
       state.profile_picture_usage = payload;
 
@@ -244,16 +245,16 @@ const mypageStore = {
       state.playlist = payload;
       console.log("재생 목록 업데이트 성공 뿌뿌");
     },
-    setVideoId(state, payload) {
-      state.videoId = payload;
-      console.log("비디오 아이디 전달");
-    },
-    setDuration(state, payload) {
-      const numbersOnly = payload.match(/\d+/g).map(Number);
-      const sec = numbersOnly[0] * 60 + numbersOnly[1];
-      state.duration = sec;
-      console.log("비디오 길이 전달", state.duration);
-    },
+    // setVideoId(state, payload) {
+    //   state.videoId = payload;
+    //   console.log("비디오 아이디 전달");
+    // },
+    // setDuration(state, payload) {
+    //   const numbersOnly = payload.match(/\d+/g).map(Number);
+    //   const sec = numbersOnly[0] * 60 + numbersOnly[1];
+    //   state.duration = sec;
+    //   console.log("비디오 길이 전달", state.duration);
+    // },
   },
   computed: {
     ...mapState({
